move test cases from error/diagnostic/tests.rs to error/tests.rs
